requirements.txt
envoy
flask
storyscript
docker
pyyaml
requests
